+ [#22792] Add function KunenaForumMessageHelper::recount(): update catid on messages ^ [#22792] KunenaForumTopic::move(): improve parameter checks + [#22792] Rethink basic move logic to handle all use cases (move, split, merge, split & merge) # [#22792] Change category/topic deletion logic on polls + [#22792] KunenaForumTopic::update(): move repeating logic into updatePostInfo() ^ [#22792] KunenaForumTopic::move(): various code simplifications, reordering + [#22792] Add functions to move() and merge() into KunenaForumTopicUserHelper ^ [#22792] Redo moving topic/posts into another category for simplicity ^ [#22792] Redo moving topic/posts into other topic to get rid of extra code + [#23442] Optimize move, split and merge for speed by using simpler and faster SQL updates + [#23442] Merging topics now updates also ownership, subscriptions and favorites
git-svn-id: file:///Users/oratzes/svn/kunena@4373 daf2aa89-cb5b-0410-b8ee-fe93ebc9b27d

+++ trunk/2.0/administrator/components/com_kunena/libraries/forum/topic/topic.php
@@ -342,43 +342,147 @@ class KunenaForumTopic extends JObject {
 		return $this->_exists;
 	}
 
-	public function move($target, $ids=false, $subject='', $shadow=false) {
+	public function move($target, $ids=false, $shadow=false, $subject='', $subjectall=false) {
+		// Warning: logic in this function is very complicated and even with full understanding its easy to miss some details!
+
+		// Cleanup input
+		if (!($ids instanceof JDate)) {
+			if (!is_array($ids)) {
+				$ids = explode(',', (string)$ids);
+			}
+			$mesids = array();
+			foreach ($ids as $id) {
+				$mesids[(int) $id] = (int) $id;
+			}
+			unset ($mesids[0]);
+			$ids = implode(',', $mesids);
+		}
+		$subject = (string) $subject;
+
+		// First we need to check if there will be messages left in the old topic
+		if ($ids) {
+			$query = new KunenaDatabaseQuery();
+			$query->select('COUNT(*)')->from('#__kunena_messages')->where("thread={$this->id}");
+
+			if ($ids instanceof JDate) {
+				// All older messages will remain (including unapproved, deleted)
+				$query->where("time<{$ids->toUnix()}");
+			} else {
+				// All messages that were not selected will remain
+				$query->where("id NOT IN ({$ids})");
+			}
+			$this->_db->setQuery ( $query );
+			$oldcount = (int)$this->_db->loadResult();
+			if ($this->_db->getErrorNum () ) {
+				$this->setError($this->_db->getError());
+				return false;
+			}
+			// So are we moving the whole topic?
+			if (!$oldcount) {
+				$ids = '';
+			}
+		}
+
+		// Find out where we are moving the messages
 		if (!$target || !$target->exists()) {
 			$this->setError(JText::printf('COM_KUNENA_MODERATION_ERROR_NO_TARGET', $this->id));
 			return false;
 
 		} elseif ($target instanceof KunenaForumTopic) {
-			if ($target->id == $this->id) {
+			// Move messages into another topic (original topic will always remain, either as real one or shadow)
+
+			if ($target == $this) {
+				// We cannot move topic into itself
 				$this->setError(JText::sprintf('COM_KUNENA_MODERATION_ERROR_SAME_TARGET_THREAD', $this->id, $this->id));
 				return false;
 			}
 			if ($this->moved_id) {
-				// TODO: translate
-				$this->setError(JText::sprintf('COM_KUNENA_MODERATION_ERROR_ALREADY_SHADOW', $this->id, $this->id));
+				// Moved topic cannot be merged with another topic -- it has no posts to be moved
+				$this->setError(JText::sprintf('COM_KUNENA_MODERATION_ERROR_ALREADY_SHADOW', $this->id));
 				return false;
 			}
-			$topic = $target;
+			if ( $this->poll_id && $target->poll_id ) {
+				// We cannot currently have 2 polls in one topic -- fail
+				$this->setError(JText::_('COM_KUNENA_MODERATION_CANNOT_MOVE_TOPIC_WITH_POLL_INTO_ANOTHER_WITH_POLL'));
+				return false;
+			}
 
 		} elseif ($target instanceof KunenaForumCategory) {
+			// Move messages into category
+
 			if ( $target->parent_id == 0 ) {
+				// Section cannot have any topics
 				$this->setError(JText::_('COM_KUNENA_MODERATION_ERROR_NOT_MOVE_SECTION'));
 				return false;
 			}
-			$oldCategory = $this->getCategory();
-			$topic = $this;
-			$this->category_id = $target->id;
+
+			// Save category information for later use
+			$categoryTarget = $target;
+			$categoryFrom = $this->getCategory();
+
+			if ($shadow || $ids) {
+				// Create new topic for the moved messages
+				$target = clone $this;
+				$target->id = 0;
+			} else {
+				// If we just move into another category, we can keep using the old topic
+				$target = $this;
+			}
+			// Did user want to change subject?
+			if ($subject) $target->subject = $subject;
+			// Did user want to change category?
+			$target->category_id = $categoryTarget->id;
 
 		} else {
 			$this->setError(JText::_('COM_KUNENA_MODERATION_ERROR_WRONG_TARGET'));
 			return false;
 
 		}
-		if ( $this->poll_id && $this->id != $topic->id ) {
-			if ( $topic->poll_id ) {
-				$this->setError(JText::_('COM_KUNENA_MODERATION_CANNOT_MOVE_TOPIC_WITH_POLL_INTO_ANOTHER_WITH_POLL'));
+
+		// For now on we assume that at least one message will be moved (=authorization check was called on topic/message)
+
+		// We will soon need target topic id, so save if it doesn't exist
+		if (!$target->exists()) {
+			// FIXME: check if we can break binding: topic->id == message->id
+			$this->setError(JText::_('COM_KUNENA_MODERATION_ERROR_WRONG_TARGET'));
+			return false;
+
+			if (!$target->save()) {
+				$this->setError($target->getError());
 				return false;
 			}
-			$query = "UPDATE #__kunena_polls SET `threadid`={$this->_db->Quote($topic->id)} WHERE `threadid`={$this->_db->Quote($this->id)}";
+		}
+
+		// Move messages (set new category and topic)
+		{
+			$query = new KunenaDatabaseQuery();
+			$query->update('#__kunena_messages')->set("catid={$target->category_id}")->set("thread={$target->id}")->where("thread={$this->id}");
+			// Did we want to change subject from all the messages?
+			if ($subjectall && !empty($subject)) $query->set("subject={$this->_db->quote($subject)}");
+
+			if ($ids instanceof JDate) {
+				// Move all newer messages (includes unapproved, deleted messages)
+				$query->where("time>={$ids->toUnix()}");
+			} elseif ($ids) {
+				// Move individual messages
+				$query->where("id IN ({$ids})");
+			}
+
+			$this->_db->setQuery ( $query );
+			$this->_db->query();
+			if ($this->_db->getErrorNum () ) {
+				$this->setError($this->_db->getError());
+				return false;
+			}
+		}
+
+		// If all messages were moved into another topic, we need to move poll as well
+		if ( $this->poll_id && !$ids && $target != $this ) {
+			// Note: We may already have saved cloned target (having poll_id already in there)
+			$target->poll_id = $this->poll_id;
+			// Note: Do not remove poll from shadow: information could still be used to show icon etc
+
+			$query = "UPDATE #__kunena_polls SET `threadid`={$this->_db->Quote($target->id)} WHERE `threadid`={$this->_db->Quote($this->id)}";
 			$this->_db->setQuery ( $query );
 			$this->_db->query ();
 			if ($this->_db->getErrorNum () ) {
@@ -386,53 +490,70 @@ class KunenaForumTopic extends JObject {
 				return false;
 			}
 		}
-		$query = new KunenaDatabaseQuery();
-		$query->update('#__kunena_messages')->set("catid={$topic->category_id}")->set("thread={$topic->id}")->where("thread={$this->id}");
-		if (!empty($subject)) $query->set("subject={$this->_db->quote($subject)}");
-		if ($ids === false) {
-			// Move the whole topic
-		} elseif ($ids instanceof JDate) {
-			// Move all newer messages (includes unapproved, deleted messages)
-			$query->where("time>={$ids->toUnix()}");
+
+		if (isset($categoryTarget)) {
+			// Move topic into another category
+
+			if ($target != $this) {
+				// Leave shadow
+				$this->moved_id = $target->id;
+			}
+			// Note: We already saved possible target earlier
+			if (!$this->save()) {
+				return false;
+			}
+
+			// Update user topic information (topic, category)
+			KunenaForumTopicUserHelper::move($this, $target);
+			// Remove topic and posts from the old category
+			$categoryFrom->update($this, -1, -$this->posts);
+			// Add topic and posts into the new category
+			$categoryTarget->update($target, 1, $this->posts);
+
 		} else {
-			// Move individual messages
-			if (is_array($ids)) $ids = implode(',', $ids);
-			$query->where("id IN ({$ids})");
-		}
-		$this->_db->setQuery ( $query );
-		$this->_db->query();
-		if ($this->_db->getErrorNum () ) {
-			$this->setError($this->_db->getError());
-			return false;
-		}
-		if ($topic->id != $this->id) {
 			// Moving topic into another topic
-			$count = 0;
-			if ($ids !== false) {
-				// We don't know if we moved all the messages -- check it
-				$query = "SELECT COUNT(*) FROM #__kunena_messages WHERE `thread`={$this->_db->Quote($this->id)}";
-				$this->_db->setQuery ( $query );
-				$count = (int)$this->_db->loadResult();
-				if ($this->_db->getErrorNum () ) {
-					$this->setError($this->_db->getError());
+
+			if (!$ids) {
+				// Create shadow topic from $this
+				$this->moved_id = $target->id;
+				if (!$shadow) {
+					// Mark shadow topic as deleted
+					$this->hold=2;
+				}
+				// Save shadow topic
+				if (!$this->save()) {
 					return false;
 				}
+
+				// Add new posts, hits and attachments into the target topic
+				$target->posts += $this->posts;
+				$target->hits += $this->hits;
+				$target->attachments += $this->attachments;
+				// Update first and last post information into the target topic
+				$target->updatePostInfo($this->first_post_id, $this->first_post_time, $this->first_post_userid,
+					$this->first_post_message, $this->first_post_guest_name);
+				$target->updatePostInfo($this->last_post_id, $this->last_post_time, $this->last_post_userid,
+					$this->last_post_message, $this->last_post_guest_name);
+				// Save target topic
+				if (!$target->save()) {
+					return false;
+				}
+
+				// Update user topic information (topic, category)
+				KunenaForumTopicUserHelper::merge($this, $target);
+				// Remove topic and posts from the old category
+				$this->getCategory()->update($this, -1, -$this->posts);
+				// Add posts into the new category
+				$target->getCategory()->update($target, 0, $this->posts);
+
+			} else {
+				// Both topics have changed and we have no idea how much: force full recount
+				// TODO: we can do this faster..
+				$this->recount();
+				$target->recount();
 			}
-			// If no messages are left, mark topic as moved
-			if (!$count) {
-				$this->moved_id = $topic->id;
-				if (!$shadow) $this->hold=2;
-			}
-			$this->recount();
-			$topic->recount();
-		} else {
-			// Moving topic into another category
-			$this->save();
-			// Remove topic from the old category
-			$oldCategory->update($this, -1, -$this->posts);
-			// Add topic into the new category
-			$target->update($this, 1, $this->posts);
 		}
+
 		return true;
 	}
 
@@ -513,34 +634,68 @@ class KunenaForumTopic extends JObject {
 		}
 		$this->_exists = false;
 
-		$db = JFactory::getDBO ();
-		// Delete user topics
-		$queries[] = "DELETE FROM #__kunena_user_topics WHERE topic_id={$db->quote($this->id)}";
-		// Delete user read
-		$queries[] = "DELETE FROM #__kunena_user_read WHERE topic_id={$db->quote($this->id)}";
-		// Delete poll (users)
-		$queries[] = "DELETE FROM #__kunena_polls_users WHERE pollid={$db->quote($this->poll_id)}";
-		// Delete poll (options)
-		$queries[] = "DELETE FROM #__kunena_polls_options WHERE pollid={$db->quote($this->poll_id)}";
-		// Delete poll
-		$queries[] = "DELETE FROM #__kunena_polls WHERE id={$db->quote($this->poll_id)}";
-		// Delete thank yous
-		$queries[] = "DELETE t FROM #__kunena_thankyou AS t INNER JOIN #__kunena_messages AS m ON m.id=t.postid WHERE m.thread={$db->quote($this->id)}";
-		// Delete all messages
-		$queries[] = "DELETE m, t FROM #__kunena_messages AS m INNER JOIN #__kunena_messages_text AS t ON m.id=t.mesid WHERE m.thread={$db->quote($this->id)}";
+		// NOTE: shadow topic doesn't exist, DO NOT DELETE OR CHANGE ANY EXTERNAL INFORMATION
+		if ($this->moved_id == 0) {
+			$db = JFactory::getDBO ();
+			// Delete user topics
+			$queries[] = "DELETE FROM #__kunena_user_topics WHERE topic_id={$db->quote($this->id)}";
+			// Delete user read
+			$queries[] = "DELETE FROM #__kunena_user_read WHERE topic_id={$db->quote($this->id)}";
+			// Delete poll (users)
+			$queries[] = "DELETE FROM #__kunena_polls_users WHERE pollid={$db->quote($this->poll_id)}";
+			// Delete poll (options)
+			$queries[] = "DELETE FROM #__kunena_polls_options WHERE pollid={$db->quote($this->poll_id)}";
+			// Delete poll
+			$queries[] = "DELETE FROM #__kunena_polls WHERE id={$db->quote($this->poll_id)}";
+			// Delete thank yous
+			$queries[] = "DELETE t FROM #__kunena_thankyou AS t INNER JOIN #__kunena_messages AS m ON m.id=t.postid WHERE m.thread={$db->quote($this->id)}";
+			// Delete all messages
+			$queries[] = "DELETE m, t FROM #__kunena_messages AS m INNER JOIN #__kunena_messages_text AS t ON m.id=t.mesid WHERE m.thread={$db->quote($this->id)}";
 
-		foreach ($queries as $query) {
-			$db->setQuery($query);
-			$db->query();
-			KunenaError::checkDatabaseError ();
-		}
+			foreach ($queries as $query) {
+				$db->setQuery($query);
+				$db->query();
+				KunenaError::checkDatabaseError ();
+			}
 
-		if ($recount) {
-			KunenaUserHelper::recount();
-			KunenaForumCategoryHelper::recount();
-			KunenaForumMessageAttachmentHelper::cleanup();
+			if ($recount) {
+				KunenaUserHelper::recount();
+				KunenaForumCategoryHelper::recount();
+				KunenaForumMessageAttachmentHelper::cleanup();
+			}
 		}
 		return $result;
+	}
+
+	public function updatePostInfo($id, $time=0, $userid=0, $message='', $name='') {
+		if ($id === false) {
+			$this->first_post_id = 0;
+			$this->first_post_time = 0;
+			$this->first_post_userid = 0;
+			$this->first_post_message = '';
+			$this->first_post_guest_name = '';
+			$this->last_post_id = 0;
+			$this->last_post_time = 0;
+			$this->last_post_userid = 0;
+			$this->last_post_message = '';
+			$this->last_post_guest_name = '';
+			return;
+		}
+		if (!$this->first_post_time || $this->first_post_time >= $time) {
+			$this->first_post_id = $id;
+			$this->first_post_time = $time;
+			$this->first_post_userid = $userid;
+			$this->first_post_message = $message;
+			$this->first_post_guest_name = $name;
+
+		}
+		if ($this->last_post_time <= $time) {
+			$this->last_post_id = $id;
+			$this->last_post_time = $time;
+			$this->last_post_userid = $userid;
+			$this->last_post_message = $message;
+			$this->last_post_guest_name = $name;
+		}
 	}
 
 	public function update($message=null, $postdelta=0) {
@@ -554,20 +709,7 @@ class KunenaForumTopic extends JObject {
 		}
 		if ($exists && $message->thread == $this->id && $message->hold == $this->hold) {
 			// If message belongs into this topic and has same state, we may need to update cache
-			if (!$this->first_post_time || $this->first_post_time >= $message->time) {
-				$this->first_post_id = $message->id;
-				$this->first_post_time = $message->time;
-				$this->first_post_userid = $message->userid;
-				$this->first_post_message = $message->message;
-				$this->first_post_guest_name = $message->name;
-			}
-			if ($this->last_post_time <= $message->time) {
-				$this->last_post_id = $message->id;
-				$this->last_post_time = $message->time;
-				$this->last_post_userid = $message->userid;
-				$this->last_post_message = $message->message;
-				$this->last_post_guest_name = $message->name;
-			}
+			$this->updatePostInfo($message->id, $message->time, $message->userid, $message->message, $message->name);
 		} elseif (!$this->moved_id) {
 			// If message isn't visible anymore, check if we need to update cache
 			if (!$exists || $this->first_post_id == $message->id) {
@@ -580,17 +722,9 @@ class KunenaForumTopic extends JObject {
 				KunenaError::checkDatabaseError ();
 
 				if ($first) {
-					$this->first_post_id = $first->id;
-					$this->first_post_time = $first->time;
-					$this->first_post_userid = $first->userid;
-					$this->first_post_message = $first->message;
-					$this->first_post_guest_name = $first->name;
+					$this->updatePostInfo($first->id, $first->time, $first->userid, $first->message, $first->name);
 				} else {
-					$this->first_post_id = 0;
-					$this->first_post_time = 0;
-					$this->first_post_userid = 0;
-					$this->first_post_message = '';
-					$this->first_post_guest_name = '';
+					$this->updatePostInfo(false);
 				}
 			}
 			if (!$exists || $this->last_post_id == $message->id) {
@@ -603,17 +737,9 @@ class KunenaForumTopic extends JObject {
 				KunenaError::checkDatabaseError ();
 
 				if ($last) {
-					$this->last_post_id = $last->id;
-					$this->last_post_time = $last->time;
-					$this->last_post_userid = $last->userid;
-					$this->last_post_message = $last->message;
-					$this->last_post_guest_name = $last->name;
+					$this->updatePostInfo($last->id, $last->time, $last->userid, $last->message, $last->name);
 				} else {
-					$this->last_post_id = 0;
-					$this->last_post_time = 0;
-					$this->last_post_userid = 0;
-					$this->last_post_message = '';
-					$this->last_post_guest_name = '';
+					$this->updatePostInfo(false);
 				}
 			}
 			if ($this->hold != KunenaForum::TOPIC_DELETED && (!$this->first_post_id || !$this->last_post_id)) {
